start rename of citycoin to citycoin-logic (errors)
