+++ contracts/test-city-wallet.clar
@@ -5,6 +5,6 @@
 (define-public (set-city-wallet (new-wallet principal))
   (begin
     (asserts! (is-eq contract-caller CONTRACT_OWNER) (err ERR_UNAUTHORIZED))
-    (contract-call? 'ST1HTBVD3JG9C05J7HBJTHGR0GGW7KXW28M5JS8QE.citycoin set-city-wallet new-wallet)
+    (contract-call? 'ST1HTBVD3JG9C05J7HBJTHGR0GGW7KXW28M5JS8QE.citycoin-logic-v1 set-city-wallet new-wallet)
   )
 )
